INTRODUCED MOTOR DRIVER SKELETONS
In this change we introduced the motor driver class and a child class
for this based on the arduino motor controller. The motor class
has two functions, setSpeed and getSpeed to control the motor. The
arduino class has an additional __init__ which allows you to pass
the motor ID.

The main Robot class got a few small updates integrating the arduino
motor class