UPDATED PROJECT LAYOUT
I've updated the project layout based on a blog post by Jean-Paul
Alderone at
http://as.ynchrono.us/2007/12/filesystem-structure-of-python-project_21.html.

In addition to this I've added a TODO.md, setup.py and __init__.py file
to make it easier to build the application. This is the first time I've
worked with setup.py in my own project so it might not work as expected
yet. At the moment there are no external dependencies yet so the file is
very basic.