INTRODUCED HARDWARE CONTROL SKELETON CLASS
This commit introduced a new generic class that will provide an
interface towards the low level hardware. At the moment the plan is to
use a serial interface between the raspberry pi and the arduino. A child
class has been created for this. This will allow the raspberry pi to send
control messages and receive data from the arduino. Have to create a
standard for sending messages between the two. At the moment a command
message will be built as follows

module moduleID command

An example could be as follows

Motor 1 speed 255

The arduino serial hardware control is dependant on the pyserial library
so this was added to setup.py. I've also made some small corrections on
the DOCSTRING format following the PEP style recommendations