CONNECTED ARDUINO SENSOR SHIELD AND DISTANCE SENSOR
Received my order from http://www.hobbycomponents.com/ so started to test the Arduino code I've written before. Still have to get a battery pack for powering the DC Motors but installed the Mega Sensor Shield onto the Arduino Mega and connected the HR-SR04 Distance Sensor. To control the Distance Sensor I'm using an external library called newPing which is also maintained on github. I've imported the module as a submodule in the arduino/externals folder.

There were quite a few syntax errors in the previous Arduino code which should be fixed. The code now is able to initialise the sensor and receive distance in cm. Also updated the motor_driver classes and it should now be able to control the Dual H-Bridge Motor Controllers. The driver initializes both the front and rear motor together so you can change the speed and direction of the motors in one go. The motor driver will have to be properly tested as soon as I've got the hardware up and running.

The main robot class will initialise all four motors and drive forward full speed. If it sees an object <= TOO_CLOSE (currently 10cm) it will stop the motors.