IMPLEMENTED SERIAL COMM. PROTOCOL FOR RASPBERRY PI
The communication protocol uses flag bytes with byte stuffing. This
means a message is marked with a character indicating the beginning and
ending of a frame. This could cause a problem when the actual message
you want to send includes the same character as the flag. To avoid this
we add an additional escape flag before these bytes to tell the other
end to treat is as part of the message. This is called byte stuffing.

The message itself consists of the following fields:
    messageID:      A numeric ID unique to the message
    module:         The module the message is intended for (e.g MOTOR)
    commandType:    The type of command (e.g START, RESTART, STOP)
    dataLen:        The lenght of the message data payload
    data:           The data payload (e.g. holding the returned Distance
                    sensor data in cm)
    checksum:       The CRC32 checksum we use to ensure data integrity

The Arduino will have to follow the same protocol to interpred and send
messages. This still has to be implemented