STARTED WITH SERIAL COMMUNICATION PROTO ON ARDUINO
Added sendMessage function to arduino raspberry_control.h which converts a message struct into a byte string. It escapes special characters and appends FRAME_FLAG to the beginning and end. Had temporary code in place that send the message out on the serial monitor as hex and seems to be working ok. Need to check efficiency as we're doing several memcpy, memmoves etc. Next stage will be to add the CRC32 checksum calculation to the message and test it against the Raspberry Pi.

I've also changed the message data field from a string to an int and removed the dataLen variable. Amended this on the raspberry hardware_controller.py as well but needs testing/checking.